+++ src/ssd/ssd_titlebar.c
@@ -3,14 +3,14 @@
 #define _POSIX_C_SOURCE 200809L
 #include <assert.h>
 #include <string.h>
-#include "labwc.h"
-#include "ssd.h"
-#include "theme.h"
-#include "common/font.h"
 #include "common/mem.h"
 #include "common/scaled_font_buffer.h"
 #include "common/scene-helpers.h"
+#include "labwc.h"
 #include "node.h"
+#include "ssd.h"
+#include "theme.h"
+#include "view.h"
 
 #define FOR_EACH_STATE(view, tmp) FOR_EACH(tmp, \
 	&(view)->ssd.titlebar.active, \
